nrf52820_hic: Extract I2C code from i2c_commands.
