+++ source/board/microbitv2/i2c_commands.h
@@ -22,6 +22,7 @@
 #ifndef I2C_COMMANDS_H
 #define I2C_COMMANDS_H
 
+#include <stdint.h>
 #include <stdbool.h>
 
 #include "cmsis_compiler.h"
@@ -33,6 +34,7 @@ typedef int32_t status_t;
 extern "C" {
 #endif
 
+/* 7-bit addresses */
 #define I2C_SLAVE_NRF_KL_COMMS      (0x70U)
 #define I2C_SLAVE_HID               (0x71U)
 #define I2C_SLAVE_FLASH             (0x72U)
@@ -175,26 +177,7 @@ typedef __PACKED_STRUCT i2cFlashCmd_tag {
     } cmdData;
 } i2cFlashCmd_t;
 
-/*! i2c Write Callback prototype */
-typedef void (*i2cWriteCallback_t)
-(
-    uint8_t*    pData,
-    uint8_t     size
-);
-
-/*! i2c Read Callback prototype */
-typedef void (*i2cReadCallback_t)
-(
-    uint8_t*    pData,
-    uint8_t     size
-);
-
-void i2c_initialize(void);
-void i2c_deinitialize(void);
-status_t i2c_registerWriteCallback(i2cWriteCallback_t writeCallback, uint8_t slaveAddress);
-status_t i2c_registerReadCallback(i2cReadCallback_t readCallback, uint8_t slaveAddress);
-void i2c_clearBuffer(void); 
-void i2c_fillBuffer(uint8_t* data, uint32_t position, uint32_t size);
+void i2c_cmds_init(void);
 
 #ifdef __cplusplus
 }
